Move helper method to utils
